Extract character map construction into CharMapBuilder
The cmap-building block was the largest remaining logical unit
inside Main.mainProcess() — roughly 100 lines covering charset
strategy selection, --lang file reading, weird/control-char
filtering, an AWT Font.canDisplay() pass, and serialization of
the surviving codepoints. It had three entry points (preset,
lang, all) plus the populateCharset helper tacked on afterwards.
Moving it into a dedicated builder lets Main.mainProcess() call
CharMapBuilder.build() in a single line and keeps the filtering
rules (including the sorted WEIRD_CHARS array) co-located with
the code that consults them.

- Add CharMapBuilder with a single public static build() method
  that takes the config, the font file path, and the cmap
  FileHandle, and returns the written cmap length in characters.
- Move the populateCharset() helper from Main into CharMapBuilder
  as a private static method.
- Lift the weird/control-char list out of the method body and
  into a CharMapBuilder.WEIRD_CHARS constant.
- Replace the ~100-line block in Main.mainProcess() with a single
  CharMapBuilder.build(...) call.
- Drop the now-unused java.util.Arrays and
  java.awt.Font.TRUETYPE_FONT imports from Main.java (plus the
  redundant java.lang.StringBuilder import).

Pure refactor: charset strategy precedence, ASCII/Latin/Cyrillic/
Greek/Latin-Ext ranges, lang-file baseline, displayable-char
filtering, stdout progress output, and the written cmap format
are all unchanged.
